Cambio de proveedor de localización de la medición
