--- src/lib/NewsSlider.svelte
@@ -1,7 +1,0 @@
-<script>
-  let contador = 0
-</script>
-
-<p>{contador}</p>
-<button on:click={() => contador ++ }>Click</button>
-
