+++ src/lib/firebase/firestore.js
@@ -23,12 +23,9 @@ const db = getFirestore();
 
 export function getUserData({ uid, city, count = 1 }) {
   return new Promise(async (res, rej) => {
-    console.log('primera bien');
-    
     const userDataRef = collection(db, 'user-data', uid, city);
-
+    
     const q = query(userDataRef, orderBy('createdAt', 'desc'), limit(count));
-
     const docs = await getDocs(q).catch((err) => rej({ success: false, err, data: null }));
 
     res({ success: true, data: docs.docs.map((doc) => doc.data()) });
