+++ src/lib/variables.js
@@ -1,4 +1,5 @@
 export const variables = {
   API_GEO_KEY: import.meta.env.VITE_API_GEO_KEY,
   API_IPINFO: import.meta.env.VITE_API_IPINFO,
+  API_LOCATIONIQ: import.meta.env.VITE_API_LOCATIONIQ,
 };
